+++ server/src/api/modules/players/services/DeletePlayerService.ts
@@ -6,7 +6,7 @@ import { setCachedPlayerId, standardize } from '../player.utils';
 
 const inputSchema = z
   .object({
-    id: z.number().positive().optional(),
+    id: z.number().int().positive().optional(),
     username: z.string().optional()
   })
   .refine(s => s.id || s.username, {
